stub랑 rdh랑 합치기 밑준비

+++ inc/qn.h
@@ -723,7 +723,7 @@ qvt_name(qnGam)
 typedef struct qnGam
 {
 	qvt_name(qnGam)*	vt;
-	volatile int		ref;
+	volatile intptr_t	ref;
 	pointer_t			ptr;
 } qnGam;
